Move main to gui.app
