+++ README.md
@@ -5,5 +5,5 @@ Property Tycoon is a game, develop for in a software engineering course.
 The game is a board game like _Monopoly_.
 
 ## License
- 
+
 MIT
