+++ src/test/java/com/watson/propert/tycoon/gui/AppTest.java
@@ -21,10 +21,11 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.watson.propert.tycoon;
+package com.watson.propert.tycoon.gui;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+import com.watson.propert.tycoon.gui.App;
 import org.junit.jupiter.api.Test;
 
 class AppTest {
